test (#185) : move cancel commute test
